+++ tests/Providers/ResponseBuilderServiceProvider.php
@@ -44,4 +44,50 @@ class ResponseBuilderServiceProvider extends \MarcinOrlowski\ResponseBuilder\Res
 
 		$this->loadTranslationsFrom(__DIR__ . '/../../src/lang', 'response-builder');
 	}
+
+	/** *******************************************************************************************
+	 * Support for multi-dimensional config array. Built-in config merge only supports flat arrays.
+	 *
+	 */
+
+	/**
+	 * Merge the given configuration with the existing configuration.
+	 *
+	 * @param string $path
+	 * @param string $key
+	 *
+	 * @return void
+	 */
+	protected function mergeConfigFrom($path, $key)
+	{
+		$config = $this->app['config']->get($key, []);
+		$this->app['config']->set($key, $this->mergeConfig(require $path, $config));
+	}
+
+	/**
+	 * Merges the configs together and takes multi-dimensional arrays into account.
+	 *
+	 * @param array $original
+	 * @param array $merging
+	 *
+	 * @return array
+	 */
+	protected function mergeConfig(array $original, array $merging)
+	{
+		$array = array_merge($original, $merging);
+		foreach ($original as $key => $value) {
+			if (!is_array($value)) {
+				continue;
+			}
+			if (!Arr::exists($merging, $key)) {
+				continue;
+			}
+			if (is_numeric($key)) {
+				continue;
+			}
+			$array[ $key ] = $this->mergeConfig($value, $merging[ $key ]);
+		}
+
+		return $array;
+	}
 }